实现项目对序列化的注入

+++ example-common/src/main/java/com/anranruozhu/common/service/UserService.java
@@ -19,6 +19,6 @@ public interface UserService {
      * @return
      */
     default  short getNumber(){
-        return 1;
+        return (short) 1;
     }
 }
